Revert "RESTRICT AUTOMERGE Set the container field of aconfig flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/32152214-9d8d-4b81-8196-5cf4ba4cb8d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibcc1a02f8cbcfc7e6605d47be5db326c3f749fe6