Revert "RESTRICT AUTOMERGE Set the container field of aconfig flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e7784154-82a7-4f66-bad2-ffc546d237d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2cf931a76b0fd9dfdcb0f7877f032c8e1ff6c1be