Revert "RESTRICT AUTOMERGE Set the container field of aconfig flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/22f03eaa-f3da-45dd-90b9-2b23a177873e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I744cd808c775bf6bcff81da06322297861e49d0d