Make stack dumping errors display better.
Move the check whether a process is cached until after the
process might have been skipped. This avoids having a lot of incorrect
cached process pids that should have been skipped.

For any pids that would have been dumped but are bypassed or have
an error, make the header include the timestamp and the command line.

For an unwind that fails:
- If it's a java process, try a native unwind.
- Display the time it took for the failed unwind.
- Check if the process is now cached and display that error message.

Bug: 339473422

Test: Trigger bugreport and verify everything continues to work.
Test: Fake an unwind failure and verify the output.
Test: Fake an unwind failure and fake a cached process and verify the output.
Test: Fake an unwind failure for a java thread, and do a native unwind.
Test: Ran the bugreports from above through the bugreport parser.
Change-Id: I3082f2ac619e179621ad3ff39f2e839d000b2576