Remove unused rust_bindgen include_dirs
In libinput_bindgen, `include_dirs` is unused (and the path does not
exist as the value is expected to be relative to the module directory).

This was a no-op before as the prop was only used for rust_ffi, but
with new build system changes where this property is used for rlib
variants, this throws an error as the path doesn't exist.

Remove the unused property from the module definition.

Bug: 254469782
Test: m
Change-Id: I27f27878117595a1f5673f0e5168bb3dd249adeb