libbinder: export fewer symbols
This change sets the default symbol visibility to hidden and then
explicitly marks symbols in the code with default visibility if they
should be part of the libbinder ABI.

We are still overexporting symbols a bit. When `LIBBINDER_EXPORTED` is
applied to a class, it exports all of that class's symbols. I've tried
to only export class members, but in a few cases it didn't seem worth
the trouble (e.g. where all the members need to be exported) or it
caused some issues (e.g. there doesn't seem to be a way to export a
class's typeinfo, which is needed for BpBinder and BpRefBase on RTTI
enabled builds).

Additionally, we are exporting many symbols only needed by binder test
binaries. Since those tests run against the standard libbinder.so, we
need to include the symbols as part of the ABI.

On aarch64, this reduces the size of libbinder_on_trusty_mock.so by
10.7% and libbinder.so by 7.5%, primarily in the .text, .dynstr, and
.dynsym sections.

Bug: 338458975
Test: TH
Change-Id: Idc348c2b12124a8bbcebeea49920267a995367b6