libbinder: reduce number exported symbols
This change sets the default symbol visibility to hidden and then
explicitly marks symbols in the code with default visibility if they
should be part of the libbinder ABI.

We are still overexporting symbols. When `LIBBINDER_EXPORTED` is applied
to a class, it exports all of that class's symbols. In later changes we
can experiment with more fine grained exports of class symbols and get
further binary size improvements.

Additionally, we are exporting many symbols only needed by binder test
binaries. Since those tests run against the standard libbinder.so, we
need to include the symbols as part of the ABI.

On aarch64, this reduces the size of libbinder_on_trusty_mock.so by 3.3%
and libbinder.so by 3.5%, primarily in the .text and .dynstr sections.

Bug: 338458975
Test: TH
Change-Id: Idc348c2b12124a8bbcebeea49920267a995367b6