Revert "Fix ping pong progress in the Kawase implementation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0551825-e13b-4b3e-8105-eb639e599318). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340625902
Change-Id: I44e401e8d6f75a719b9fe6f4c0dd021781f358f1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com