Revert "Fix typo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f7524a28-c179-4e47-936a-d6507719c35b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 341141037
Change-Id: If9c65c43b859fea88a956b0560c6ac67854b3c70
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com