simplify Parcel object position sorting
Ensure mObjects is always sorted and simplify the code to make use of
that fact.

Saves about 10kb in binary size for libbinder.so.

Writing an object to the middle of parcel was already unsupported but we
didn't check for it. Now it will be an error if it would break the
object sorting invariant (so that we don't need to do an insertion
sort).

Bug: 236990334, 338458975
Test: TH
Change-Id: Ic8616cb734530feb96d4a8318e484d4020aa0494