Revert "Add addService to pass in dump priority flags."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c8204d51-fc16-4e6e-b8af-ad80aa3e5373). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 345331905 Broken build 11933593 on git_main on husky-next-user
Change-Id: I6c320b637f397899563f86c8e8d6c95203f10c5a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com