Revert "Add addService to pass in dump priority flags."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8dc84ab1-c9f7-4020-b390-090bf6261361). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 345335119
Change-Id: I6e6ea1bcec3c8608f8513a886be0f285fd4471eb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com