Add a flag to set the same refresh rate for both virtual and physical screens
In the 3D scene of AR glasses, a virtual screen should be added, and its
refresh rate should match that of the AR glasses. Currently, the virtual
screen is encountering significant frame drops.

To address this issue, set the refresh rate of the virtual screen to be
the same as the physical screen by including a flag.

Change-Id: If14b5876407d365ddacb98f857457bcdc9a997ec
Signed-off-by: Jia Jia <jia.jia@zte.com.cn>