Revert "Add perfetto traces to ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c164a68d-8ff8-4fee-9923-9cdf1bf58742). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I23a3a0a6ce5fcc550a03963434a629be29a2a2f4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com