Revert "Add perfetto traces to ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0feebed4-f7f0-4b92-8843-de5254893bb6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7536876cfaa97a924bac00776abd23d656a0ac81
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com