Revert "Add perfetto traces to ServiceManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/59ff6b39-de1d-4e84-b303-39b0d8a67384). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc37a377e6dcf31852309379e286b8e9bb0aa48c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com