Revert "libc++fs is part of libc++ now."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c61a0655-aa8c-47ac-befc-93c8afe9271e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 350527370
Change-Id: Iaf6ed2af2b1b05fd2b1d0f09151b7459f09f7922
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com