Revert "Define android.software.credentials.xml as prebuilt_etc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50a70453-b4b3-441c-92f8-db6856b01d41). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c672b69bbfdbe333c970b8828ca1f84dacae75f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com