Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d1df9d8-efe8-4df7-8232-b945830636e5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36f93a7dd32daf94b3444f41e99d7b8aaf37b347
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com