Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bd8308a0-2228-4af4-a7f5-d60776a50115). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I628d469022befddf96c6b04c1e674775f1166d6c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com