Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1db01fb-5029-44fa-ac7d-a18f5ce9c15b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I45fdff598cfeebfaaf0092659b26eedc05395d3f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com