Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9dde5bba-66dc-4e6d-8faf-7d0d4a7bec6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a65af0dc57339606b968e34f7eebc44bb5c96e7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com