Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5bfc4c57-ece8-4602-9e54-5265e0ce0bda). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1795152e9802ac58f1a0e7c61d7f2ddc0c92ae0b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com