Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66ba84df-1be1-4fa1-a22e-b2f9b575d655). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icafb7943f74a2d722a0de255fe9c1b5f30e42107
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com