Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f913cff6-a3bf-4546-8852-14bfa03c5db2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17d60c1815299b00fb2c835b61ca78bea2fe7cf3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com