Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/352e67af-a346-4ff1-8ccc-ecbe495483ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I64c9124ce12057a748e2a3f326c5ec13357a6f03
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com