Revert "Placeholder for binder freeze notification API."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ed4a201-0ff5-46b4-8494-d2ca718fe446). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib4e76625f23ae0e62de57a185028a912fcb57c3d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com