Updating barrerFrameNumber when the ExternalTexture is null but frame has changed.
If ExternalTexture is null, and Layer::setBuffer() is called with a
null buffer, the framenumber of bufferData will not update to
mDrawingState, and if the next buffer is fulled with a buffer barrier,
the transaction queue will be stalled since
mDrawingState.barrierFrameNumber is less than
bufferData.barrierFrameNumber.

it most likely occurs that client does setBuffer -> cacheBuffers(by
IPC delivery) -> unsetBuffer -> setBuffer, and reuse the same buffer
in both of two setBuffer, if the first setBuffer cached buffer
successfully, and the second setBuffer will do cache-hit, but SF did
not do anything for this buffer, finally it will cause SF get a null
ExternalTexture instance from ClientCache.

So we should update barrierFrameNumber in setBuffer() when buffer is
null but frame number has changed, we also need to clear barrier via
flushState.bufferLayersReadyToPresent when the transaction with
buffer but without ExternalTexture is applied if needed.

Bug: 354116298

Test: presubmit

Change-Id: I232dbb19ca1fe0b15a52e915c139d23a5682182b
Signed-off-by: jiayongqiang <jiayongqiang@xiaomi.com>