Updating flushState.bufferLayersReadyToPresent when the ExternalTexture is null.
If ExternalTexture is null, and Layer::setBuffer() is called with a
null buffer, the framenumber of bufferData will not update to
mDrawingState, and if the next buffer is fulled with a buffer barrier,
the transaction queue will be stalled since
mDrawingState.barrierFrameNumber is less than
bufferData.barrierFrameNumber.

So we should clear barrier via flushState.bufferLayersReadyToPresent
when the transaction with buffer but without ExternalTexture is
applyed.

Bug: 354116298

Test: presubmit

Change-Id: I232dbb19ca1fe0b15a52e915c139d23a5682182b
Signed-off-by: jiayongqiang <jiayongqiang@xiaomi.com>