Revert "Support IAccessor in libbinder for RPC services"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cbac161-b81c-495d-9fab-0215156ae502). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If7cd436e3c22f8579b2e185eeb9300b5f55c2186
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com