Revert "Support IAccessor in libbinder for RPC services"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0161ffb5-e1c0-4226-aad2-ffb1c195fdd5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b9af669ec53972069773477a1c389b4788fccbc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com