EGL Multifile Blobcache: Update size on key reuse
Before this fix, if a key was submitted with a different value, we
would increase the size of the cache by the new amount. Some
platforms are reusing keys a lot and running out of memory.

Now we check to see if it is contained, then remove it before adding
the new value.

This also updates removeFromHotCache to be non-blocking on pending
work. All places that need to preserve existing cache contents
already wait. For this scenario, we don't need to wait, allowing
any pending writes to disk to complete in order, resulting in the
final correct entry being last.

Test: libEGL_test, EGL_test, ANGLE trace tests, apps
Bug: b/351867582
Change-Id: I66fe9cde18e468e541a80296c80f2234ac61acb0