BPF: remove duplicate cflags and btf
With the addition of the -Wall, -Werror, -Wextra cflags to the bpf soong
module as default, the are no longer necessary in the bpf definition in
Android.bp files.

Additionally, btf:true is now set by default. This should be the default
for all bpf programs unless the are mainlined and targeting legacy
versions of bpfloader.

Test: treehugger
Change-Id: Ic5029319e82a45cdbfa6c5f4c4375dab7fea7284
Signed-off-by: Neill Kapron <nkapron@google.com>