Revert "Support IAccessor in libbinder for RPC services"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3346909e-2d5c-4d25-8955-c871ebd41d6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd8154b062f4447642f48ed86a220d2261454b09
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com