Relax ACLs for kprobe-related files in tracing sysfs
Among the /sys/kernel/tracing files that are already writable,
kprobes-related files and folders were not. This would prevent anyone
except root to read or write these files.
This patch extends the ACL permissions of kprobe event files with read
and write permissions to everyone, aiming at allowing traced (part of
Perfetto) and shell to use this kernel tracing feature. This extension
is limited to userdebug builds.

Bug: 359542775
Test: traced succeeds in creating and activating kprobe events
Change-Id: I6aa13e34987a5bf89a5a96b41d64dbb498aca0ae
Signed-off-by: Alessio Balsini <balsini@google.com>