Revert "Support IAccessor in libbinder for RPC services"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f644da5e-a601-44b8-aa92-873cda2e15ff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I776453539b2715fb3a0dd84a850082d99d56eb75
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com