Add a new service manager for local RPC binder services
This is used to allow the use of service manager APIs in systems
without kernel binder or a servicemanager process that is that is set
the system-wide context 0 binder object.

This also adds a new multi-service service manager that handles both
the calls to (1) the kernel binder service manager process at
context 0 for the system and (2) a new service manager implementation
for the RPC binder services.

For some methods like getService(), we will first check the kernel
binder service manager process if it's installed and then check with the
RPC binder service manager if the service wasn't found.

Test: atest binderRpcTest vm_accessor_test
Bug: 358427181

Change-Id: Ibef52415bf1509cdcd6c9cb65dbb8e20ec08d007