Move getService below addService.
A recent change to the aidl interface (https://r.android.com/3194614)
triggered a bug in Pre-reboot Dexopt. The root cause is that `artd` in
chroot links against the new libbinder_ndk.so in chroot, which talks to
servicemanager outside of chroot through an unstable aidl interface.
The changes to the unstable aidl interface caused the communication to
be broken.

The problem and the solutions are discussed in details in b/356858364.

As a long term solution, we changed Pre-reboot Dexopt to bind-mount
/system{,_ext}/lib{,64} from the old system (see
https://r.android.com/q/topic:pr-dexopt-bind-mount-libs). However, this
solution requires an ART change and an sepolicy change on the source
system, so it only benefits OTA from Android 16+ to future versions.

Meanwhile, we are also exploring the possibility to avoid the infinate
loop in `waitForService` if the service cannot be started, to prevent a
failure in starting the artd_pre_reboot service from blocking OTA.

For OTA from Android 15 to Android 16, the short-term solution is to
make the addService API the same as before and keep it stable until
Android 16 is finalized. Hence, we need this CL.

Bug: 356858364
Test: OTA from 24Q3 to (ToT + this CL)
Change-Id: If72b3e19187ad9e5e497567fa66a35319147ddd7