Revert "Make binder available to com.android.microfuchsia APEX"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1c0223f-4063-4c41-bafe-014da5351387). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361108146
Change-Id: Ic0af2e9b128eaf5df07f73ffe6f596c83bed5b5a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com