Revert "remove libbinder_ndk_helper_headers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/acdffb91-388d-4598-947f-9bc2ef45e740). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362363320
Change-Id: I35810dd667bfb7c5a672b491d6d5147820d9bf68
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com