Revert "remove libbinder_ndk_helper_headers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8352082b-2ea3-4430-97c5-78717d553d5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362378410
Change-Id: I9bf3cd61901eff4b5ce6d184904a0a4ed703ed8c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com