Remove the redundant parameters in applyTransactionsLocked
In the major Android version upgrade from 13 to 14, the addCommittedTransactions function in the applyTransactionsLocked function that requires the vsyncId parameter has been moved to the updateLayerSnapshots and updateLayerSnapshotsLegacy functions. The rest of the operations in the applyTransactionsLocked function no longer require the vsyncId parameter, so the declaration and definition information about vsyncId in the applyTransactionsLocked function can be removed here.

Bug：362415544

Change-Id: Ic4ded171e90bcfb7a0d58df2c272f4c219be3a04
Merged-In: <change-id-of-cherry-pick>
Signed-off-by: xiangshuangxi@xiaomi.corp-partner.google.com