Revert "remove libbinder_ndk_helper_headers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7f33b2ea-990a-4a19-9510-0e1f132cae8f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362378980
Change-Id: Id1f2bb671e0307b2e029dd058de78096bddc46b0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com