Revert "remove libbinder_ndk_helper_headers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7441f00b-78ac-4d5f-a8a4-75bfbeebb8bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362379072
Change-Id: I69b4dd440b8ba7040f37b6d5f9cbdd13a4780e05
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com