Revert "remove libbinder_ndk_helper_headers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/721ae29d-1faa-4201-b176-e7d0fe9a6a38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362378366
Change-Id: I4706bea6aadfdad153529a06a1094af43a43666a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com