Revert "remove libbinder_ndk_helper_headers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c37a0246-4ed3-48d9-8bb9-0435a456c54d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362362251
Change-Id: I213dace25924ede5b81a8573a2ac065d99bf72a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com