Revert "Move tracing calls to libbinder_ndk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/703e843e-5fd6-4005-9821-23dd5cf2623e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 367851525
Change-Id: I0312170edddd45256f8b1449fdd80f9dae623951
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com