Revert "Move tracing calls to libbinder_ndk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27d4a7e3-69e7-4df9-9636-2a35a9c2e51f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 367851662
Change-Id: I4e901e365fd303326373f8797e6d3143f88f7c98
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com