Remove LL-NDK details from API_LEVEL_AT_LEAST.
The LL-NDK half of this check is only necessary when
ro.vendor.api_level is less than 202404, which isn't the case for any
ToT builds. Since that's unnecessary, we can remove the LL-NDK details
here.

This is a step towards being able to put API_LEVEL_AT_LEAST in the NDK
proper, since the macro in the NDK will not have the LL-NDK details in
it.

Bug: None
Test: None
Change-Id: Iafb8f8422a66a4cf15c09a4a7002c41df38bcbac