Revert "Add methods to convert between HardwareBuffer and native..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1484171a-80cc-4f53-92d3-33c635d4b9e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I877cd97e88b1598235d7451be08dd7ead86b9997
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com