Revert "Add Rust wrapper around native_handle_t."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1484171a-80cc-4f53-92d3-33c635d4b9e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I314fe4f98f921d2b387f5531ec1c080bc2dc0a0c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com