Revert "Add methods to create a new NativeHandle and get fds and..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1bfa4557-9c7e-40dd-832f-28664afdf8a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fcd78ba318fbffddd30a7a4a149a67a33e80cc8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com