Revert "Add methods to create a new NativeHandle and get fds and..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89663408-11a9-4aa1-a3bb-74a9d7dd4055). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie9a577fad854728efca8d449a527a7dd88c6ab23
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com