Revert "Add methods to create a new NativeHandle and get fds and..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c84b27a-3d01-41fe-b5bc-38d61cddc8e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icca88366379b03d08626b5da64a5fd058ae417c7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com