Revert "Revert "DO_NOT_MERGE Fix primaryRangeIsSingleRate + touc..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/450ceb80-a9a7-4995-a84c-63a9556937d2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 369700807
Change-Id: Ia50d83fed6ec0df6ba35d11dc6b3b7d7a7e7a682
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com