Revert "Revert "DO_NOT_MERGE Fix primaryRangeIsSingleRate + touc..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/083f30db-f19c-4801-b146-4f4e81689596). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 369701147
Change-Id: I96d9cea4fbe39d20bc396c459d49ac906dbb3a6c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com