Revert "Trusty: update LL-NDK mock."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ff2b96b-9371-42dc-b520-1df63974da41). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I267c98474cef31d08d13be08d87f2bd2c4a52e71
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com