Side-loaded ANGLE "isolated" instead of "shared isolated"
ANGLE library names in the system partion and in the apk are the same
(e.g. libGLESv2_angle.so). When GLES is used as the system driver the
libaries in the system partition are not loaded so this doesn't cause
issues. However, when ANGLE is installed as the system driver
(such as on cuttlefish, or enabled via persist.graphics.egl=angle),
the system partition libraries can get preloaded into zygote
(SkiaGL without ro.zygote.disable_gl_preload=true) and then
unload_system_driver() can't actually unload the libs. With "shared"
namespace, when ANGLE is configured to be side-loaded from an apk, this
results in the system driver being loaded unexpectedly.

"isolated" namespaces require additional setup to enable loading
.so dependencies (including core libs like libdl.so). This is already
implemented by linkDriverNamespaceLocked for the "updatable gfx driver"
which is loaded "isolated". Reusing that for side-loaded apk fixes the
issue on cuttlefish.

Bug: b/371356548
Tested: cvd create --gpu_mode=guest_swiftshader, side-load ANGLE apk
Change-Id: Ib5dc9157b93f9a96a42e69fb134139d698ddefc2