Revert "libbinder: statusToString for status_t errors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aee87614-426e-4512-9040-0c3a387e20d2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72acafa8c659971d2a4404ef77169a27c5272060
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com