Revert "libbinder: statusToString for status_t errors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/88cb7990-93e9-4bff-b87c-c4e3c40e8786). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d84b255cbb56fce9997e1dd9a406e3e53f5d198
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com