Revert "libbinder: statusToString for status_t errors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c5e131a8-a7d7-43fe-a34c-bf8dd372b06b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75778f5a3cc22f3a3ac2ba735cfafb5ea62af132
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com