Revert "Add Tracing for Binder Caching"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1c32a38f-f9c8-40fc-9d3d-f137eab68cd5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iccc4c2e10020fef709a389bbca68ff7e3d747b88
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com