Complete the trace information LLM to LayerLifecycleManager
When analyzing the trace of performance issues, the LLM-related information in the LLM:commitChanges time consumption is unclear to newcomers or non-professionals. It is recommended to complete LLM with LayerLifecycleManager, which can make it clear to the person analyzing the trace and reduce the time loss of searching in the source code.

Signed-off-by:
Shuangxi Xiang <xiangshuangxi@xiaomi.corp-partner.google.com>

Change-Id: I5ed05d0f5aa4c5e98fbd9b5275f3db39ccb76fed