Revert "Add more const to 'instances' in registerAccessorProvider"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5872ee9a-9f63-4bea-9f58-172da97292a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfbd6d14dd7d0c6d65e7d03bcd2c7b6503f12b02
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com