Revert "Add default maxBufferCount in GraphicBufferProducer"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b67be47c-74bf-4e37-b35f-77a54b808046). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic67939c1abdada9e55352770433f7d0acd845691
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com