Fix the problem that NavigationBar doesn't disappear when locking the screen
On Android V, sometimes NavigationBar still appears when it is on the
locked screen. When framework applyTransactions in
native, sometimes there are both hide navigation bar and show navigation
bar behavior in a transaction, so that
NavigationBar doesn't disappear when locking the screen. So if it has a
hide navigation bar behavior in a transaction,
it won't show a navigation bar.

Change-Id: Ia47a2ed1fb9bad8a60e77e25dc484b3d213a898c
Signed-off-by: Jia Jia <jia.jia@zte.com.cn>