Revert "Define VINTF fragment as a module type"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1af91a92-178e-4b0e-9d2f-af8872c1e470). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb3cfa814e7cfd1d1a7d9744eb9a3fd0d7300d69
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com