Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1c21271-8c9f-434f-b1fb-98b5376e797a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ea620b9a0711a0da3b7bf2abe51320b078db6f4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com