Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4b2221a7-eb30-44ba-898c-d685c7bb5664). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2320f9ebf0754bd1f031b7d4e1566aefa081b917
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com