Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/30476ee5-1bef-44ac-9448-0b489761ad5f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48bb82ea72125c92ce27b0bcdedfc7d02d97daf6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com