Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f7b72bd4-a1d6-4384-824a-3df2e5340112). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I310f8ee5ce8101aadb6694e650a81d9cb58d96cf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com