Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ccf9e4a-7915-4830-87fb-19f7dbd8574a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia0fd116f19c268f9d78a51f2960b9090165e3a1b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com