Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3edda1a3-a356-412d-870a-31192e00bf86). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia35700ad086b7130bfa007a91200d1487110ceb4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com