Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/704aeb77-7f02-4223-a427-1f4a05f9df60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I38f69653b65d41315ebfb36eac2d58372596bb49
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com