Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87c62a94-562d-4944-b1e9-2cddc63934e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2470b5d55f62c7f1fb2e71dcc295bb3372918204
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com