Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6c66f65-4923-4f1e-9512-5e521ba7e187). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2cc9ba15716b198b2cdf151179d2db166ce15e60
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com