Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6087d6b8-c4c1-4bb2-866a-36e57ca93424). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c2fb91d855545ca3e1e2af6e2871a076df5645f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com