Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/adbb7e34-9983-4869-a1cb-68a54cfe5761). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic87447b4cf108eab62b0fd0b25865f92b22b003a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com