Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/146b614f-019d-4b6c-b472-9b239641c9bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8bd50f56e7a5372cf8e1c7d2ddca1440a2359602
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com