Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4cfa6db6-5c21-4f2a-9ba5-381009aee98f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3266d8da45c2b5648ad35f3dc23f1f208b94762f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com