Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34099500-0e44-4c98-8ec5-50119eacf960). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I334401173739206d52cbdb21033a72ff42064ede
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com