Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce87a0ec-22b4-48ed-9f90-07f78616629c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id364f615b02ed517e46f457d1dd7c0f9690c1230
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com