Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c4c8675-264b-4f40-af45-3c5e0221749a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic2b707d318a67d2e49aad73f2a78ac2c20b6a9f2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com