Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0f62f3ab-bf4d-4f13-a695-347a9e2b0b30). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifb135ca775ba481931d2b63386bfd522fdc3fd42
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com