Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9e3b163-09cb-46e8-ac8a-1fc450a23de3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d88f186273764a651c783848371cef3b39cd858
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com