Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dae0a3ec-6e0e-4591-b3aa-19afc192a2b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7839554f710cde3bef1d79ee749c777d6363fa22
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com