Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf819dbd-908f-44f6-808a-6e38f3770ed7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I921800c7cd06db919b87d1baa92d3cd98d6fba68
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com