Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9dcee292-c192-4ba2-861f-bfb199683f12). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e5027fb0d2385b9e83402877be4f9c5331621fa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com