Fix the issue of generating a layer cycle in LayerHierarchy when a layer is reparented to its descendants.
When a layer reparented, if the parent layer is its child in the layer tree, a layer circle will be formed.

When CaptureLayers, surfaceflinger will check whether the layer is secure. The layer cycle will cause a recursive loop and eventually cause surfaceflinger to crash.

Therefore, we need to use the 'hasLayerCycle' method to determine whether the reparent operation will form a layer circle. If so, reparent should not be executed.

Bug: 382620055

Test: manually test

1. new SurfaceControl1
2. new SurfaceControl2.reparent(SurfaceControl1)
3. new SurfaceControl3.reparent(SurfaceControl2)
4. SurfaceControl1.reparent(SurfaceControl3)
5. ScreenCapture.captureLayers(SurfaceControl3, new Rect(0,0,200,200), 1.0f);

Change-Id: I5565f4667e28177293880c12f2a0a3b8995d4a0a
Signed-off-by: guodalin <guodalin@xiaomi.corp-partner.google.com>