Revert "Wrap AHardwareBuffer_LockPlanes too."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b1ce52e8-1269-46a0-8538-d72a4ffc4715). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf0d087d41671ad659e3ee878935a6a002df776b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com