Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46016dd7-ba0a-49b2-995d-bbee739b9d9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383169523
Change-Id: Iab49a7f6a3eb7bf5b9b51d66eb458da5e6574b45
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com