Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63969c1a-47e1-4cca-ad75-f3230aff9d62). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383169208
Change-Id: I4d9fa46cde0e0f4a2ed9f140abcde8eb0a105268
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com