Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c1686f02-33c3-4c54-bfbf-e860bb2090cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383168926
Change-Id: I2c11d02efa6f8fd9f436337fe2a877948c64e4ca
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com