Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03a74637-47c3-4908-b496-dd5a3c2970c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383199096
Change-Id: Iac038551b940a380e7f489953a0d12dc98e84ab9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com