Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e97fc68-df88-4544-8cdd-08d8de0ba478). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383199870
Change-Id: Ibdc3be5d2d55d1c805a0de09b566f963c92df617
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com